refactor(core): use gql.tada on blog and category queries
